requirements.txt
asgiref==3.10.0
attrs==25.4.0
Django==5.2.7
django-autoslug==1.9.9
django-cities-light==3.10.2
django-cors-headers==4.9.0
django-debug-toolbar==6.0.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Markdown==3.9
mysqlclient==2.2.7
pillow==12.0.0
progressbar2==4.5.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-utils==3.9.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
Unidecode==1.4.0
uritemplate==4.2.0
